refactor: move InformerConfigHolder to appropriate package
